Move schema creation class into `mysql/schema_creation.rb`
Current master branch includes many schema creation improvements in
MySQL. It extract these features to the appropriate file.
